@@ -10,14 +10,15 @@
         }
     ],
     "require": {
-        "php": ">=5.4",
-        "studio-42/elfinder": "^2.1.10",
-        "league/flysystem": "^1.0",
-        "league/flysystem-cached-adapter": "^1.0",
-        "intervention/image": "^2.0"
+        "php": ">=8.0.2",
+        "intervention/image": "^2.0",
+        "league/flysystem": "^3.0",
+        "league/glide": "^2.2",
+        "studio-42/elfinder": "^2.1.58"
     },
     "suggest": {
-        "league/glide": "1.x to display images through Glide"
+        "league/glide": "2.x to display images through Glide",
+        "spatie/flysystem-dropbox": "2.x to use Dropbox as filesystem"
     },
     "autoload": {
          "classmap": ["elFinderVolumeFlysystem.php"],
